% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_QC_comparison_plots.R
\name{create_QC_comparison_plots}
\alias{create_QC_comparison_plots}
\title{Comparison Plots for Quality Control (QC)}
\usage{
create_QC_comparison_plots(
  df1,
  df2,
  nri_all_info,
  variables,
  variable_units,
  plot_types,
  no_of_plots_per_page,
  output_dir
)
}
\arguments{
\item{df1}{First data frame (e.g., model 1 predictions).}

\item{df2}{Second data frame (e.g., model 2 predictions).}

\item{nri_all_info}{Data frame containing additional information such as state_name, recordid2017, and year. Ensure these columns are present in the dataset.}

\item{variables}{A vector of variable names for which plots should be created. To generate lag difference plots, append "_diff" to the variable name (e.g., "somc_diff").}

\item{variable_units}{A vector of corresponding units (e.g., "(ton/ha)", "(cm)") for each variable. Note: g/m² has already been converted to ton/ha. Include brackets in units.}

\item{plot_types}{A vector specifying the plot type for each variable ("line", "scatter", or "both").}

\item{no_of_plots_per_page}{Maximum number of plots to display per page.}

\item{output_dir}{Directory where output plots will be saved.}
}
\value{
The function generates and saves line plots as a PDF and scatter plots as image files.
}
\description{
Generates comparison plots based on specified variables and plot types.
(please make sure to install ggplot2, dplyr, gridExtra install in your computer)
}
\examples{
variables <- c("somsc", "n2oflux", "strmac1", "sdrema", "egrain1", "fertot21", "annppt", "accrst")
variable_units <- c("(ton/ha)", "(ton/ha)", "(ton/ha)", "(ton/ha)", "(ton/ha)", "(ton/ha)", "(cm)", "(ton/ha)")
plot_types <- c("line", "scatter", "both", "scatter", "scatter", "scatter", "scatter", "scatter")
output_dir <- "C:/temp/DayCent_output_comparison"
highest_no_of_plots_per_page <- 6
df_491 <- daycent_output_from_491_version_model
df_513 <- daycent_output_from_513_version_model
create_plots(df1 = df_491, df2 = df_513, nri_all_info = nri_all_info, 
             variables = variables, variable_units = variable_units, 
             plot_types = plot_types, no_of_plots_per_page = highest_no_of_plots_per_page, 
             output_dir = output_dir)

}
